+++ scout/server/blueprints/institutes/static/variants_list_scripts.js
@@ -1,20 +1,27 @@
 $(document).ready(function() {
-  $('#variants_table').DataTable( {
+  $('#variants_table').DataTable({
       paging: false,
-			searching: true,
-			layout: {
-				topStart: 'buttons',
-				topEnd: {
-            search: {
-                placeholder: 'Search'
-            }
-        }
-			},
+      searching: true,
+      layout: {
+          topStart: 'buttons',
+          topEnd: {
+              search: {
+                  placeholder: 'Search'
+              }
+          }
+      },
       buttons: [
           'copyHtml5',
           'excelHtml5',
+          {
+              text: 'Managed Variants infile',
+              className: 'btn btn-primary',
+              action: function () {
+                  window.location.href = exportAsManagedUrl;
+              }
+          }
       ]
-  } );
+  });
 
   $('[data-toggle="tooltip"]').tooltip();
-} );
+});